% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d2plnr}
\alias{d2plnr}
\title{Data example 1}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 800 rows and 25 columns.
}
\usage{
d2plnr
}
\description{
Data example 1
}
\keyword{datasets}
