% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{familyrisk}
\alias{familyrisk}
\title{Family risks of alcohol use}
\format{
An object of class \code{data.frame} with 2724 rows and 16 columns.
}
\usage{
familyrisk
}
\description{
Family risks of alcohol use
}
\keyword{datasets}
