% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hlt_start.R
\name{get_hlt_start}
\alias{get_hlt_start}
\title{hlt Starting Values}
\usage{
get_hlt_start(x, nchains = 1)
}
\arguments{
\item{x}{hlt model fit object}

\item{nchains}{number of chains to get starting values}
}
\value{
a list of lists with starting values for each chain
}
\description{
Get starting values from hlt fit object
}
