% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inirt.R
\name{inirt}
\alias{inirt}
\title{Explanatory Item Response Theory}
\usage{
inirt(data, model, predictors = NULL, dims, method = c("vb", "hmc"), ...)
}
\arguments{
\item{data}{data frame}

\item{model}{text specification of the model. Not yet implemented. Instead use the other arguments}

\item{predictors}{list of predictors}

\item{dims}{dimensions}

\item{method}{Choose estimation method. Choices are method = 'vb' (default) for variational Bayes, or method = 'hmc' for Hamiltonian Monte Carlo.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Explanatory Item Response Theory
}
