% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument.R
\name{instrument}
\alias{instrument}
\title{Bayesian Explanatory Multidimensional Item Response Theory}
\usage{
instrument(
  data,
  model,
  itype = "2pl",
  exploratory = TRUE,
  method = "hmc",
  fweights = NULL,
  chains = 1,
  ...
)
}
\arguments{
\item{data}{a named \code{data.frame} or \code{matrix}}

\item{model}{text specification of the model. See examples or website.}

\item{itype}{item type of the items. Options are '1pl', '2pl', '3pl'}

\item{exploratory}{fit exploratory MIRT model? Only used with multidimensional
model that is not higher-order or bifactor}

\item{method}{Choose estimation method. Use method = 'hmc' for Hamiltonian
Monte Carlo. Other option is method = 'vb' for Variational Bayes. HMC should
be the default for published results. VB is experimental.}

\item{fweights}{frequency weights. These weights represent how many times
each row of the data should be repeated in the analysis. fweights weight
the likelihood function.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}

\item{ranef_id}{id for which items belong to the current random effect.}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Fit a multidimensional item response theory model using Stan.
See github.com/theta2pack/doc for examples, tutorials, and documentation.
}
