% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings.R
\name{loadings}
\alias{loadings}
\title{Loadings from a thata2 model fit object}
\usage{
loadings(x, rotate = c("varimax", "promax"), ...)
}
\arguments{
\item{x}{model fit object from theta2, i.e., x = theta2(data, ...)}

\item{rotate}{what rotation to use? Default is 'varimax'. Other option is
'promax'. Only one can be specified at a time}

\item{...}{other arguments to pass down to the varimax and promax functions
in the stats package. See ?varimax for more information.}
}
\value{
a list containing the unrotated loadings, rotated loadings according
to the rotate argument, and the matrix of posterior estimates of the unrotated
loading parameters.
}
\description{
Get factor loadings (rotated and unrotated) from a theta2 model fit object.
}
