% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_chains.R
\name{merge_chains}
\alias{merge_chains}
\title{Merge Chains from hlt method}
\usage{
merge_chains(x, ...)
}
\arguments{
\item{x}{object of class "hltObjList"}

\item{...}{other arguments}
}
\value{
a list of class `hltObj`. This class constructs a single `hltObj`
from a list of model fits by merging the chains into one matrix of draws.
}
\description{
Merge Chains from hlt method
}
