% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model.R
\name{parse_model}
\alias{parse_model}
\title{Parse inirt model}
\usage{
parse_model(model, data, exploratory = FALSE)
}
\arguments{
\item{model}{text string which gives model definition}

\item{data}{data frame passed from user}

\item{exploratory}{do exploratory multidimensional IRT? TRUE/FALSE

This function takes a user inputted model fron inirt::inirt and the data,
translates the input and returns the data which stan and inirt::inirt require to
fit the appropriate model.}
}
\value{
list containing translated data to set up and fit the model
}
\description{
Parse inirt model. For internal use only.
}
