% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_regression_eq.R
\name{parse_regression_eq}
\alias{parse_regression_eq}
\title{Parse Regression Equation for inirt Model}
\usage{
parse_regression_eq(model, data)
}
\arguments{
\item{model}{string specifying regression formula}

\item{data}{data frame}
}
\value{
a list with the named elements:
data = new data frame coded for fitting the regression model given the formula
predictors = numeric column number for the fixed effects
predictors_ranef = numeric column number for the independent random effects
ranef_id = numeric ID for which random effect variance each random effect belongs
predictors_ranef_cor = numeric column number for the correlated random effects
n_pranef_cor = vector length of correlated random effects
}
\description{
Parse an inirt model regression equation. For internal use only.
This function is called by the inirt::inirt() function to parse
a user specified regression equation. See examples on this page.
}
\examples{
data = as.data.frame(matrix(0, 10, 20))
names(data) = paste0("x", 1:20)
data$School = paste0("s", rep(1:5, each = 2))
data$age = runif(10, 10, 20)

reg_data = parse_regression_eq(model = "t1 ~ (1|School) + (age|School) + x12 + x13 + x15", data = data)
reg_data = parse_regression_eq(model = "t1 ~ (1 + age|School) + x12 + x13 + x15", data = data)
reg_data = parse_regression_eq(model = "t1 ~ (1 + age|School)", data = data)
reg_data = parse_regression_eq(model = "t1 ~ 0", data = data)
reg_data = parse_regression_eq(model = "t1 ~ x12 + x13 + x15", data = data)



}
