% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_theta_eq.R
\name{parse_theta_eq}
\alias{parse_theta_eq}
\title{Parse Theta Equation for inirt Model}
\usage{
parse_theta_eq(model, data, exploratory = FALSE)
}
\arguments{
\item{model}{string specifying regression formula}

\item{data}{data frame}

\item{exploratory}{do exploratory multidimensional IRT? TRUE/FALSE}
}
\value{
a list with the named elements:
}
\description{
Parse an inirt model theta equation. For internal use only.
This function is called by the inirt::inirt() function to parse
a user specified theta equation. See examples on this page.
}
\examples{

# data = as.data.frame(matrix(0, 10, 50))
names(data) = paste0("x", 1:50)
data$School = paste0("s", rep(1:5, each = 2))
data$age = runif(10, 10, 20)
mod_theta = "c(1:10, 20:35) + x12 + x13 + x14"
  mod_theta = "x12 + x13 + x14"
  mod_theta = "c(1:10, 20:35)"
mod = "t1 = c(1:10, 20:35) + x12 + x13 + x14"
mod = "t1 = c(1:10, 20:35)"
mod = "t1 = c(1:10)"
mod = "t1 = x12 + x13 + x14"
mod = "t1 = x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8"
mod = "t1 = x1 + c(20:35)"
mod = c("t1 = x1 + c(20:35)", "t2 = c(1:10, 20:35)", "t3 = x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8")
mod = c("t1 = x1 + c(20:35)", "t2 = c(1:10, 20:35)", 
  "t3 = x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8", "t4 = t1 + t2 + t3")
parse_regression_eq(mod, data = data)
mod = c("t1 = x1 + c(20:35)", "t2 = c(1:10, 20:35)", "t3 = x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8")
parse_regression_eq(mod, data = data)
mod = c("t1 = c(1:50)", "t2 = c(1:50)", "t3 = c(1:50)", "t4 = c(1:50)")
parse_regression_eq(mod, data = data, exploratory = TRUE)





}
