% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta2-package.R
\docType{package}
\name{theta2-package}
\alias{theta2-package}
\alias{theta2}
\title{The 'theta2' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.7. https://mc-stan.org
}
